Migrate quick sort
- Move implementation to header file
- Add missing curly braces
- Update README
- Delete old quick_sort.cpp
- Fix naming conflict issue in merge_sort.hpp
